+++ src/compiler/dyvil/tools/compiler/ast/method/MethodMatch.java
@@ -5,9 +5,9 @@ import dyvil.tools.compiler.util.Util;
 public final class MethodMatch implements Comparable<MethodMatch>
 {
 	public final IMethod	method;
-	public final int		match;
+	public final float		match;
 	
-	public MethodMatch(IMethod method, int match)
+	public MethodMatch(IMethod method, float match)
 	{
 		this.method = method;
 		this.match = match;
