+++ src/dwt/dyvil/tools/compiler/ast/dwt/DWTReference.java
@@ -47,7 +47,7 @@ public class DWTReference extends ASTNode implements IValue
 	}
 	
 	@Override
-	public int getTypeMatch(IType type)
+	public float getTypeMatch(IType type)
 	{
 		return 0;
 	}
